:truck: `REPL` => move in dedicated folder + split in files
