+++ src/app/side-nav-tabs/bottom/index.ts
@@ -1,1 +1,1 @@
-export * from './repl.tab'
+export * from './notebook'
