+++ src/app/side-nav-tabs/bottom/notebook/index.ts
@@ -1,0 +1,2 @@
+export * from './repl.tab'
+export * from './cell-javascript'
